requirements.txt
pytest==7.4.0
selene==2.0.0rc9
selenium==4.15.0
allure-pytest==2.13.0
pytest-xdist==3.3.0
webdriver-manager==4.0.0
python-dotenv==1.0.0